feat(server): separate files into directories
