+++ server/src/auth/controllers/auth.controller.ts
@@ -1,6 +1,6 @@
-import { Controller, Request, Post, UseGuards, Req } from '@nestjs/common';
-import { AuthService } from './auth.service';
-import { LocalAuthGuard } from './guards/local-auth.guard';
+import { Controller, Request, Post, UseGuards } from '@nestjs/common';
+import { AuthService } from '../services/auth.service';
+import { LocalAuthGuard } from '../guards/local-auth.guard';
 
 @Controller('auth')
 export class AuthController {
