+++ server/src/auth/strategy/local.strategy.ts
@@ -1,7 +1,7 @@
 import { PassportStrategy } from '@nestjs/passport';
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { Strategy } from 'passport-local';
-import { AuthService } from '../auth.service';
+import { AuthService } from '../services/auth.service';
 import { ErrorCodes } from 'src/common/enum/error-codes.enum';
 
 @Injectable()
