+++ server/src/gates/gates.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { NodesClientModule } from 'src/nodes-client/nodes-client.module';
 import { UsersClientModule } from 'src/users-client/users-client.module';
-import { GatesService } from './gates.service';
-import { GatesController } from './gates.controller';
+import { GatesService } from './services/gates.service';
+import { GatesController } from './controllers/gates.controller';
 import { MqttModule } from 'src/mqtt/mqtt.module';
 
 @Module({
