+++ server/src/nodes-client/nodes-client.service.ts
@@ -7,10 +7,10 @@ import { instanceToPlain, plainToInstance } from 'class-transformer';
 import mongoose from 'mongoose';
 import { ErrorCodes } from 'src/common/enum/error-codes.enum';
 import { NodeDto } from 'src/nodes/dtos/node.dto';
-import { NodesService } from 'src/nodes/nodes.service';
+import { NodesService } from 'src/nodes/services/nodes.service';
 import { PublicUserDto } from 'src/users-client/dtos/public-user.dto';
 import { UserDto } from 'src/users/dtos/user.dto';
-import { UsersService } from 'src/users/users.service';
+import { UsersService } from 'src/users/services/users.service';
 import { CreateNodeRequestDto } from './dtos/create-node-request.dto';
 import { GetNodesResponseDto } from './dtos/get-nodes-response.dto';
 import { UpdateNodeRequestDto } from './dtos/update-node-request.dto';
