+++ server/src/users/users.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 import { User, UserSchema } from './schemas/user.schema';
-import { UsersService } from './users.service';
+import { UsersService } from './services/users.service';
 
 @Module({
   imports: [
